Changes to the website after testing
I decided to make changes to the code after testing the website.
I would most likely need to interact with the Database engine separated from this project due to the bugs in the website.
I am trying to figure out the cause of the bugs and get rid of them.

+++ Bahrainouna Project/quiztime.php
@@ -4,7 +4,7 @@
     <title>Creating Form</title>
     <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/@picocss/pico@2/css/pico.min.css">
 </head>
-<img src="bahrain.jpg" alt="Bahrainouna Q&A" width="500px" height="150px" style="align-self: center; align-items: center;"><br>
+<img src="bahrain.jpg" alt="Bahrainouna Q&A" width="500px" height="150px" style="margin-left: 35%;"><br>
 <body id="formBody554">
     <h3 align="center">Choose an option below to start action in this website:</h3><br>
     <nav class="container">
@@ -21,7 +21,7 @@
         <?php
             try {
                 $selectionMade = new PDO("mysql:host=localhost;port=3306;dbname=theChallenge", "root", "MOH123ha");
-                $ownerPDO = $selectionMade->query("SELECT ID, Name FROM Form");
+                $ownerPDO = $selectionMade->query("SELECT * FROM Form");
                 foreach($ownerPDO as $own):
         ?>
                     <input type="radio" name="selectForm" id="pOwnest<?= $own["ID"];?>" value="<?=$own["ID"];?>">
@@ -43,21 +43,36 @@
     <input class="container" type="submit" name="submission1" id="pSubmit100" value="Confirm">
 </form><br>
 <?php
+    $showQuestions = null; $retrieved = null;
     try {
+        $shower = null;
+        if (!isset($_SESSION) || !isset($_SESSION["cpr"])){
+            session_start();
+            $sessionPDO = new PDO("mysql:host=localhost;port=3306;dbname=theChallenge", "root", "MOH123ha");
+            $sessionInserter = $sessionPDO->query("SELECT * FROM Login");
+            $counter = $sessionInserter->rowCount();
+            foreach ($sessionInserter as $mySession){
+                $counting = 0;
+                if ($counting == ($counter-1)){
+                    $_SESSION["cpr"] = $mySession["CPR"];
+                }
+            }
+        }
         $justDoPDO = new PDO("mysql:host=localhost;port=3306;dbname=theChallenge", "root", "MOH123ha");
-        if (isset($_GET["selectForm"]) && isset($_GET["submission1"])){
-            $showQuestions = $justDoPDO->prepare("SELECT ID, Question FROM QandA WHERE OwnerForm = :owner6");
-            $showQuestions->bindParam(':owner6', $_GET["selectForm"]);
-            $showQuestions->execute();
+        if (isset($_GET["selectForm"]) && isset($_GET["submission1"]) && $_GET["submission1"] == "Confirm"){
+            $showQuestions = $justDoPDO->prepare("SELECT * FROM QandA WHERE OwnerForm = ?");
+            $showQuestions->execute([$_GET["selectForm"]]);
             $shower = $showQuestions->fetch();
             foreach ($shower as $realing):
-                $showAnswers = $justDoPDO->prepare("SELECT theKey, Details, FROM Answers WHERE OwnerForm = ? AND ReferQA = ?");
-                $showAnswers->execute([$_GET["selectForm"], $realing["ID"]]);
-                $retrieved = $showAnswers->fetch();
-                echo $realing["ID"].". ".$realing["Question"]."<br>";
-                foreach ($retrieved as $display):
-                    echo $display["theKey"].". ".$display["Details"]."<br>";
-                endforeach;
+                if ($realing != null || $shower != null):
+                    $showAnswers = $justDoPDO->prepare("SELECT theKey, Detail FROM Answers WHERE OwnerForm = ? AND ReferQA = ?");
+                    $showAnswers->execute([$_GET["selectForm"], $realing["ID"]]);
+                    $retrieved = $showAnswers->fetch();
+                    echo $realing["ID"].". ".$realing["Question"]."<br>";
+                    foreach ($retrieved as $display):
+                        echo $display["theKey"].". ".$display["Detail"]."<br>";
+                    endforeach;
+                endif;
 ?>
                 <form action="quiztime.php" method="post">
                     <label for="labeling">Answer: </label>
@@ -66,7 +81,7 @@
                 </form><br>
 <?php
                 try {
-                    if (isset($_POST["labeling"]) && isset($_POST["submission2"])){
+                    if (isset($_POST["labeling"]) && isset($_POST["submission2"]) && $retrieved != null){
                         $keyClass = new TheKeys($_POST["labeling"], $_GET["selectForm"]);
                         $keyPDO = $justDoPDO->prepare("INSERT INTO theKeys VALUES (?,?)");
                         $keyPDO->execute([$keyClass->getTheLetter(), $keyClass->getOwnerForm()]);
@@ -111,57 +126,68 @@
             <script>alert("Thank you for trying to answer the form!");</script>
 <?php
         }
-        $keyCollected = $justDoPDO->prepare("SELECT * FROM theKeys WHERE OwnerForm = :owner7");
-        $keyCollected->bindParam(':owner7', $_GET["selectForm"]);
-        $keyCollected->execute();
-        $keysDisplays = $keyCollected->fetch();
-        $questionsCollected = $justDoPDO->prepare("SELECT * FROM QandA WHERE OwnerForm = :owner8");
-        $questionsCollected->bindParam(':owner8', $_GET["selectForm"]);
-        $questionsCollected->execute();
-        $questionsDisplays = $questionsCollected->fetch();
-        $answersCollected = $justDoPDO->prepare("SELECT * FROM Answers WHERE OwnerForm = ? AND ReferQA = ?");
-        $answersCollected->execute([$_GET["selectForm"], $questionsDisplays["ID"]]);
-        $answersDisplays = $answersCollected->fetch();
-        $correctCollected = $justDoPDO->prepare("SELECT * FROM Corrections WHERE answers = ? AND questions = ? AND OwnerForm = ?");
-        $correctCollected->execute([$answersDisplays["ID"], $questionsDisplays["ID"], $_GET["selectForm"]]);
-        $correctDisplays = $correctCollected->fetch();
-        if (isset($keysDisplays) && isset($questionsDisplays) && isset($answersDisplays) && isset($correctDisplays)):
-            $founderQ = ""; $founderA = ""; $founderK = ""; $countGot = 0; $countTotal = 0;
-            foreach ($correctDisplays as $corrected):
-                foreach ($questionsDisplays as $questioned){
-                    if ($corrected["questions"] == $questioned["ID"]){
-                        $founderQ = $questioned;
-                        break;
-                    }
-                }
-                foreach ($answersDisplays as $answered){
-                    foreach ($keysDisplays as $keyD){
-                        if ($keyD["theLetter"] == $answered["theKey"]){
-                            $founderK = $keyD["theLetter"];
-                            break;
-                        }
-                    }
-                    if ($answered["ID"] == $corrected["answers"]){
-                        $founderA = $answered;
-                        break;
-                    }
-                }
-                if ($founderQ != "" && $founderA != "" && $founderK != ""){
-                    if ($corrected["answer"] == $founderA["ID"] && $corrected["question"] == $founderQ["ID"]){
-                        if ($founderA["theKey"] == $founderK):
-                            $countGot = $countGot + 1;
-                        endif;
-                    }
-                }
-            endforeach;
-            echo "<br>Your score = ".$countGo."/".$countTotal."<br>";
-        else:
+        $keysTotal = $justDoPDO->query("SELECT * FROM theKeys");
+        $questionsTotal = $justDoPDO->query("SELECT * FROM QandA");
+        if ($keysTotal->rowCount() >= $questionsTotal->rowCount()-1):
+            if (isset($_GET["selectForm"]) && isset($_GET["submission1"]) && $shower != null && $showQuestions != null):
+                $keyCollected = $justDoPDO->prepare("SELECT * FROM theKeys WHERE OwnerForm = ?");
+                $keyCollected->execute([$_GET["selectForm"]]);
+                $keysDisplays = $keyCollected->fetch();
+                $questionsCollected = $justDoPDO->prepare("SELECT * FROM QandA WHERE OwnerForm = ?");
+                $questionsCollected->execute([$_GET["selectForm"]]);
+                $questionsDisplays = $questionsCollected->fetch();
+                $answersCollected = null; $answersDisplays = null; $correctCollected = null; $correctDisplays = null; 
+                if ($questionsDisplays["ID"] != null):
+                    $answersCollected = $justDoPDO->prepare("SELECT * FROM Answers WHERE OwnerForm = ? AND ReferQA = ?");
+                    $answersCollected->execute([$_GET["selectForm"], $questionsDisplays["ID"]]);
+                    $answersDisplays = $answersCollected->fetch();
+                    if ($answersDisplays["ID"] != null):
+                        $correctCollected = $justDoPDO->prepare("SELECT * FROM Corrections WHERE answers = ? AND questions = ? AND OwnerForm = ?");
+                        $correctCollected->execute([$answersDisplays["ID"], $questionsDisplays["ID"], $_GET["selectForm"]]);
+                        $correctDisplays = $correctCollected->fetch();
+                    endif;
+                endif;
+                if (isset($keysDisplays) && isset($questionsDisplays) && isset($answersDisplays) && isset($correctDisplays)):
+                    if ($answersDisplays != null && $correctDisplays != null):
+                        $founderQ = ""; $founderA = ""; $founderK = ""; $countGot = 0; $countTotal = 0;
+                        foreach ($correctDisplays as $corrected):
+                            foreach ($questionsDisplays as $questioned){
+                                if ($corrected["questions"] == $questioned["ID"]){
+                                    $founderQ = $questioned;
+                                    break;
+                                }
+                            }
+                            foreach ($answersDisplays as $answered){
+                                foreach ($keysDisplays as $keyD){
+                                    if ($keyD["theLetter"] == $answered["theKey"]){
+                                        $founderK = $keyD["theLetter"];
+                                        break;
+                                    }
+                                }
+                                if ($answered["ID"] == $corrected["answers"]){
+                                    $founderA = $answered;
+                                    break;
+                                }
+                            }
+                            if ($founderQ != "" && $founderA != "" && $founderK != ""){
+                                if ($corrected["answer"] == $founderA["ID"] && $corrected["question"] == $founderQ["ID"]){
+                                    if ($founderA["theKey"] == $founderK):
+                                        $countGot = $countGot + 1;
+                                    endif;
+                                }
+                            }
+                        endforeach;
+                        echo "<br>Your score = ".$countGo."/".$countTotal."<br>";
+                    else:
 ?>
-            <script>
-                alert("Your answers have not been corrected.");
-                console.log("Database couldn't be accessed for correcting answers.");
-            </script>
+                        <script>
+                            alert("Your answers have not been corrected.");
+                            console.log("Database couldn't be accessed for correcting answers.");
+                        </script>
 <?php
+                    endif;
+                endif;
+            endif;
         endif;
     }
     catch(PDOException $e){
